Pulled up name and description into TaggedEntity.
