+++ src/main/java/gov/usds/case_issues/db/model/TaggedEntity.java
@@ -2,6 +2,7 @@ package gov.usds.case_issues.db.model;
 
 import javax.persistence.Column;
 import javax.persistence.MappedSuperclass;
+import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 
 import org.hibernate.annotations.NaturalId;
@@ -16,15 +17,29 @@ public abstract class TaggedEntity extends UpdatableEntity {
 	@JsonProperty("tag")
 	@Pattern(regexp="[-\\w]+")
 	private String externalId;
+	@NotNull
+	@Column(nullable=false)
+	private String name;
+	private String description;
 
 	protected TaggedEntity() { /* for hibernate */ }
 
-	protected TaggedEntity(String externalId) {
+	protected TaggedEntity(String externalId, @NotNull String name, String description) {
 		super();
 		this.externalId = externalId;
+		this.name = name;
+		this.description = description;
 	}
 
 	public String getExternalId() {
 		return externalId;
+	}
+
+	public String getName() {
+		return name;
+	}
+
+	public String getDescription() {
+		return description;
 	}
 }
